# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "orcid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcid"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "required": [
          "id",
          "name",
          "orcid",
          "position"
        ],
        "title": "AuthorRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_PaperSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaperSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[PaperSummary]",
        "type": "object"
      },
...
      "PaperDetail": {
        "properties": {
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract"
          },
          "arxiv_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arxiv Id"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/AuthorRef"
            },
            "title": "Authors",
            "type": "array"
          },
          "authors_complete": {
            "default": true,
            "title": "Authors Complete",
            "type": "boolean"
          },
          "citation_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Count"
          },
          "cited_by_in_graph": {
            "default": 0,
            "title": "Cited By In Graph",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_stub": {
            "default": false,
            "title": "Is Stub",
            "type": "boolean"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "metrics": {
            "$ref": "#/components/schemas/GraphMetrics"
          },
          "openalex_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openalex Id"
          },
          "publication_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "references_in_graph": {
            "default": 0,
            "title": "References In Graph",
            "type": "integer"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/TopicRef"
            },
            "title": "Topics",
            "type": "array"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "venue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VenueRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "abstract",
          "description",
          "year",
          "publication_date",
          "doi",
          "openalex_id",
          "arxiv_id",
          "url",
          "language",
          "citation_count",
          "authors_complete",
          "is_stub",
          "sources",
          "authors",
          "venue",
          "topics",
          "keywords",
          "cited_by_in_graph",
          "references_in_graph",
          "metrics"
        ],
        "title": "PaperDetail",
        "type": "object"
      },
...
      "PaperSearchResults": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/PaperHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "mode": {
            "enum": [
              "keyword",
              "semantic",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "semantic_available": {
            "default": true,
            "title": "Semantic Available",
            "type": "boolean"
          }
        },
        "required": [
          "query",
          "mode",
          "hits",
          "semantic_available",
          "note"
        ],
        "title": "PaperSearchResults",
        "type": "object"
      },
      "PaperSummary": {
        "properties": {
          "arxiv_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arxiv Id"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "citation_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Count"
          },
          "cited_by_in_graph": {
            "default": 0,
            "title": "Cited By In Graph",
            "type": "integer"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_stub": {
            "default": false,
            "title": "Is Stub",
            "type": "boolean"
          },
          "pagerank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagerank"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "year",
          "venue",
          "authors",
          "doi",
          "arxiv_id",
          "citation_count",
          "cited_by_in_graph",
          "pagerank",
          "is_stub"
        ],
        "title": "PaperSummary",
        "type": "object"
      },
...
      "SearchResults": {
        "properties": {
          "authors": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Authors",
            "type": "array"
          },
          "papers": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Papers",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "query",
          "papers",
          "authors",
          "topics"
        ],
        "title": "SearchResults",
        "type": "object"
      },
      "SimilarPaper": {
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "method": {
            "enum": [
              "coupling",
              "cocitation",
              "ppr",
              "semantic"
            ],
            "title": "Method",
            "type": "string"
          },
          "paper": {
            "$ref": "#/components/schemas/PaperSummary"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "shared": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shared"
          }
        },
        "required": [
          "paper",
          "method",
          "score",
          "shared",
          "explanation"
        ],
        "title": "SimilarPaper",
        "type": "object"
      },
...
      "TopicRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "id",
          "name",
          "score"
        ],
        "title": "TopicRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Research intelligence over a citation knowledge graph.",
    "title": "ResearchGraph API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/papers": {
      "get": {
        "operationId": "list_papers_api_papers_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 1600,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year From"
            }
          },
          {
            "in": "query",
            "name": "year_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 1600,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year To"
            }
          },
          {
            "in": "query",
            "name": "topic_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "year",
              "enum": [
                "year",
                "pagerank",
                "cited_by"
              ],
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PaperSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List papers",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/papers/{paper_id}": {
      "get": {
        "operationId": "get_paper_api_papers__paper_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Paper details",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/papers/{paper_id}/citations": {
      "get": {
        "operationId": "get_citations_api_papers__paper_id__citations_get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "title": "Paper Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PaperSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Papers citing this paper",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/papers/{paper_id}/references": {
      "get": {
        "operationId": "get_references_api_papers__paper_id__references_get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "title": "Paper Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PaperSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Papers cited by this paper",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/papers/{paper_id}/related": {
      "get": {
        "operationId": "related_papers_api_papers__paper_id__related_get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "title": "Paper Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimilarPaper"
                  },
                  "title": "Response Related Papers Api Papers  Paper Id  Related Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Related papers by Personalized PageRank over the local citation neighbourhood",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/papers/{paper_id}/semantic-similar": {
      "get": {
        "operationId": "semantically_similar_papers_api_papers__paper_id__semantic_similar_get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "title": "Paper Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimilarPaper"
                  },
                  "title": "Response Semantically Similar Papers Api Papers  Paper Id  Semantic Similar Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Papers whose title and abstract are closest in meaning (embedding similarity)",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/papers/{paper_id}/similar": {
      "get": {
        "operationId": "similar_papers_api_papers__paper_id__similar_get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "title": "Paper Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "default": "coupling",
              "enum": [
                "coupling",
                "cocitation"
              ],
              "title": "Method",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimilarPaper"
                  },
                  "title": "Response Similar Papers Api Papers  Paper Id  Similar Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Structurally similar papers (bibliographic coupling or co-citation)",
        "tags": [
          "papers"
        ]
      }
    },
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Keyword search over papers, authors and topics",
        "tags": [
          "search"
        ]
      }
    },
    "/api/search/papers": {
      "get": {
        "operationId": "search_paper_hits_api_search_papers_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "hybrid",
              "enum": [
                "keyword",
                "semantic",
                "hybrid"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSearchResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search papers by keyword (BM25), meaning (embeddings) or both (hybrid)",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}